requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
bcrypt==4.0.1
pyjwt
python-multipart
pydantic[email]
python-dotenv
